WIP: fixup files view; refactor contacts and add markup
